requirementes.txt
transformers
datasets
accelerate
bitsandbytes
trl
peft
torch
huggingface_hub
scipy
python-dotenv